LPS-47433 Reformat journal Subscription tests
