+++ portal-impl/test/integration/com/liferay/portlet/journal/service/JournalSubscriptionContainerModelTest.java
@@ -1,0 +1,70 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portlet.journal.service;
+
+import com.liferay.portal.kernel.test.ExecutionTestListeners;
+import com.liferay.portal.test.LiferayIntegrationJUnitTestRunner;
+import com.liferay.portal.test.MainServletExecutionTestListener;
+import com.liferay.portal.test.Sync;
+import com.liferay.portal.test.SynchronousMailExecutionTestListener;
+import com.liferay.portal.util.BaseSubscriptionContainerModelTestCase;
+import com.liferay.portal.util.test.RandomTestUtil;
+import com.liferay.portal.util.test.TestPropsValues;
+import com.liferay.portlet.journal.model.JournalArticle;
+import com.liferay.portlet.journal.model.JournalFolder;
+import com.liferay.portlet.journal.util.test.JournalTestUtil;
+
+import org.junit.runner.RunWith;
+
+/**
+ * @author Zsolt Berentey
+ * @author Roberto Díaz
+ */
+@ExecutionTestListeners(
+	listeners = {
+		MainServletExecutionTestListener.class,
+		SynchronousMailExecutionTestListener.class
+	})
+@RunWith(LiferayIntegrationJUnitTestRunner.class)
+@Sync
+public class JournalSubscriptionContainerModelTest
+	extends BaseSubscriptionContainerModelTestCase {
+
+	@Override
+	protected long addBaseModel(long containerModelId) throws Exception {
+		JournalArticle article = JournalTestUtil.addArticle(
+			group.getGroupId(), containerModelId);
+
+		return article.getResourcePrimKey();
+	}
+
+	@Override
+	protected long addContainerModel(long containerModelId) throws Exception {
+		JournalFolder folder = JournalTestUtil.addFolder(
+			group.getGroupId(), containerModelId,
+			RandomTestUtil.randomString());
+
+		return folder.getFolderId();
+	}
+
+	@Override
+	protected void addSubscriptionContainerModel(long containerModelId)
+		throws Exception {
+
+		JournalFolderLocalServiceUtil.subscribe(
+			TestPropsValues.getUserId(), group.getGroupId(), containerModelId);
+	}
+
+}
